Crate pwasm_ethereum

Externs library to interact with Ethereum-like network

Structs

Error
Generic wasm error

Functions

address
Get address of currently executing account

balance
Get balance of the given account

block_hash
Returns hash of the given block or H256::zero()

block_number
Get the block's number

call
Message-call into an account

call_code
Like call, but with code at the given address

coinbase
Get the current block’s beneficiary address (the current miner account address)

create
Create a new account with the given code

difficulty
Get the block's difficulty.

gas_limit
Get the block's gas limit.

input
Allocates and requests call arguments (input)

log
Creates log entry with given topics and data.

origin
Get execution origination address

read
Performs read from the storage.

ret
Sets a call return value

sender
Get caller address

static_call
Like call, but this call and any of it's subcalls are disallowed to modify any storage (will return Err(Error) in this case)

suicide
Halt execution and register account for deletion

timestamp
Get the block's timestamp

value
Get deposited value by the instruction/transaction responsible for this execution.

write
Performs write to the storage