Function pwasm_ethereum::call

pub fn call(
    gas: u64, 
    address: &Address, 
    value: U256, 
    input: &[u8], 
    result: &mut [u8]
) -> Result<(), Error>

Message-call into an account

Arguments:

gas- a gas limit for a call. A call execution will halt if call exceed this amount

address - an address of contract to send a call

value - a value in Wei to send with a call

input - a data to send with a call

result - a mutable reference to be filled with a result data

Returns:

Call is succeed if it returns Result::Ok(()) If call returns Result::Err(Error) it means tha call was failed due to execution halting